pyproject.toml
[tool.poetry]
name = "com2fun"
version = "0.1.8"
description = "Transform document into function"
authors = ["Zhengmian Hu <huzhengmian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
openai = ">=1.9.0"
anthropic = {version = ">=0.18.1", optional = true}

[tool.poetry.group.dev.dependencies]
python-dotenv = ">=1.0.0"

[tool.poetry.extras]
anthropic = ["anthropic"]
all=["anthropic"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
